SaaS/templates/service-template/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Cache
django-redis==5.4.0

# Authentication
PyJWT==2.8.0
django-allauth==0.57.0

# Utilities
python-decouple==3.8
requests==2.31.0
celery==5.3.4

# Development (remover em produção)
django-debug-toolbar==4.2.0

# Production optimizations
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
sentry-sdk==1.38.0
